+++ packages/core/src/languages/__tests__/javascript.test.ts
@@ -17,6 +17,18 @@ suite("Built-in JavaScript Language Plugin", function() {
     expect(() => new JavaScriptLanguagePlugin({})).not.to.throw();
   });
 
+  suite("::getEmbeds", function() {
+    test("does not throw", function() {
+      const file = {
+        type: "css",
+        content: ".foo { color: red; }",
+      };
+
+      const plugin = new JavaScriptLanguagePlugin();
+      expect(() => plugin.getEmbeds(file)).not.to.throw();
+    });
+  });
+
   suite("::getExpressions", function() {
     let plugin: WebManglerLanguagePlugin;
 
@@ -44,17 +56,17 @@ suite("Built-in JavaScript Language Plugin", function() {
 
     test("get languages", function() {
       const plugin = new JavaScriptLanguagePlugin();
-      const result = plugin.getLanguages();
-      expect(result).to.include.keys(DEFAULT_EXTENSIONS);
+      const result = Array.from(plugin.getLanguages());
+      expect(result).to.include.members(DEFAULT_EXTENSIONS);
     });
 
     test("get configured languages", function() {
       const jsExtensions = ["jsx", "ts"];
 
       const plugin = new JavaScriptLanguagePlugin({ jsExtensions });
-      const result = plugin.getLanguages();
-      expect(result).to.include.keys(DEFAULT_EXTENSIONS);
-      expect(result).to.include.keys(jsExtensions);
+      const result = Array.from(plugin.getLanguages());
+      expect(result).to.include.members(DEFAULT_EXTENSIONS);
+      expect(result).to.include.members(jsExtensions);
     });
   });
 });