+++ packages/core/src/languages/html/embeds/css/index.ts
@@ -1,0 +1,7 @@
+import { getStyleAttributesAsEmbeds } from "./style-attribute";
+import { getStyleTagsAsEmbeds } from "./style-tag";
+
+export default [
+  getStyleAttributesAsEmbeds,
+  getStyleTagsAsEmbeds,
+];