+++ packages/core/src/languages/html/embeds/js/common.ts
@@ -1,0 +1,6 @@
+/**
+ * The {@link WebManglerEmbed} type of embeds found by this module.
+ *
+ * @since v0.1.21
+ */
+export const EMBED_TYPE = "js";