+++ packages/core/src/manglers/__tests__/css-variables.test.ts
@@ -749,9 +749,12 @@ suite("CSS Variable Mangler", function() {
         const cssVariableMangler = new CssVariableMangler({
           reservedCssVarNames: reserved,
         });
+
         const result = cssVariableMangler.options();
         expect(result).to.have.property("reservedNames");
-        expect(result.reservedNames).to.include.members(reserved);
+
+        const reservedNames = Array.from(result.reservedNames as string[]);
+        expect(reservedNames).to.include.members(reserved);
       });
     });